Fix Gitaly build for 32-bit operating systems
https://gitlab.com/gitlab-org/gitaly/-/merge_requests/5059 removed a
cast for the platform-dependent `rusage` package. On Raspberry Pi
(armv7) platforms, these are 32-bit values, but on x86 64 they are
64-bit values. However, the `stats.RecordX` functions expect 64-bit
values.

This commit fixes the build by casting the values to 64-bit values.

Closes https://gitlab.com/gitlab-org/omnibus-gitlab/-/issues/7321

+++ internal/command/command.go
@@ -443,11 +443,12 @@ func (c *Command) logProcessComplete() {
 		stats.RecordSum("command.real_time_ms", realTime.Milliseconds())
 
 		if ok {
-			stats.RecordMax("command.maxrss", rusage.Maxrss)
-			stats.RecordSum("command.inblock", rusage.Inblock)
-			stats.RecordSum("command.oublock", rusage.Oublock)
-			stats.RecordSum("command.minflt", rusage.Minflt)
-			stats.RecordSum("command.majflt", rusage.Majflt)
+			// Cast to int64 is necessary for 32-bit platforms
+			stats.RecordMax("command.maxrss", int64(rusage.Maxrss))
+			stats.RecordSum("command.inblock", int64(rusage.Inblock))
+			stats.RecordSum("command.oublock", int64(rusage.Oublock))
+			stats.RecordSum("command.minflt", int64(rusage.Minflt))
+			stats.RecordSum("command.majflt", int64(rusage.Majflt))
 		}
 	}
